+++ package.json
@@ -19,10 +19,10 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^6.0.3",
-    "@angular/core": "^6.0.3",
-    "@angular/compiler": "^6.0.3",
     "@angular/animations": "^6.0.3",
+    "@angular/common": "^6.0.3",
+    "@angular/compiler": "^6.0.3",
+    "@angular/core": "^6.0.3",
     "@angular/forms": "^6.0.3",
     "@angular/http": "^6.0.3",
     "@angular/platform-browser": "^6.0.3",
@@ -37,32 +37,30 @@
     "applicationinsights": "^1.1.0",
     "axios": "^0.18.0",
     "cookie-parser": "^1.4.4",
+    "core-js": "^2.5.4",
+    "ejs": "^2.6.1",
     "exception-formatter": "^1.0.7",
+    "express": "^4.16.4",
+    "express-session": "^1.15.6",
     "global-tunnel-ng": "^2.7.1",
     "govuk-frontend": "2.5.1",
     "jwt-decode": "^2.2.0",
+    "log4js": "^3.0.6",
     "ngx-cookie": "4.0.2",
     "ngx-cookie-service": "^2.1.0",
     "otp": "^0.1.3",
-    "core-js": "^2.5.4",
-    "ejs": "^2.6.1",
-    "express": "^4.16.4",
-    "express-session": "^1.15.6",
-    "log4js": "^3.0.6",
     "p-iteration": "^1.1.7",
     "rxjs": "^6.0.0",
     "rxjs-compat": "^6.4.0",
     "session-file-store": "^1.2.0",
+    "smoothscroll-polyfill": "^0.4.3",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "^6.0.3",
-    "@angular-devkit/build-ng-packagr": "~0.6.8",
     "@angular-devkit/build-angular": "~0.6.8",
-    "ng-packagr": "^3.0.0-rc.2",
-    "tsickle": ">=0.25.5",
-    "tslib": "^1.7.1",
+    "@angular-devkit/build-ng-packagr": "~0.6.8",
     "@angular/cli": "~6.0.8",
+    "@angular/compiler-cli": "^6.0.3",
     "@angular/language-service": "^6.0.3",
     "@types/express": "^4.16.0",
     "@types/jasmine": "~2.8.6",
@@ -80,10 +78,13 @@
     "karma-jasmine": "1.1.2",
     "karma-jasmine-html-reporter": "0.2.2",
     "mocha": "^5.2.0",
+    "ng-packagr": "^3.0.0-rc.2",
     "ngrx-store-freeze": "^0.2.4",
     "nodemon": "^1.18.9",
     "protractor": "~5.3.0",
     "ts-node": "^5.0.1",
+    "tsickle": ">=0.25.5",
+    "tslib": "^1.7.1",
     "tslint": "~5.9.1",
     "typescript": "2.9.2"
   }
