+++ src/users/components/index.ts
@@ -1,5 +1,5 @@
-import {InviteUserFormComponent} from './invite-user-form.component';
+import {InviteUserFormComponent} from './invite-user-form/invite-user-form.component';
 
 export const components: any[] = [InviteUserFormComponent]
 
-export * from './invite-user-form.component';
+export * from './invite-user-form/invite-user-form.component';
