+++ src/users/containers/userform/user-form.component.ts
@@ -18,7 +18,8 @@ export class UserFormComponent implements OnInit {
   isSubmitted = false;
   inviteUserForm: FormGroup;
   formValidationErrors$: Observable<any>;
-  formValidationErrorsArray$: Observable<any>;
+  formValidationErrorsArray$: Observable<string[]>;
+  //
   errorMessages = {
     firstName: 'Enter first name',
     lastName: 'Enter last name',
@@ -30,6 +31,7 @@ export class UserFormComponent implements OnInit {
   ngOnInit(): void {
     this.formValidationErrors$ = this.store.pipe(select(fromStore.getGetInviteUserList));
     this.formValidationErrorsArray$ = this.store.pipe(select(fromStore.getGetInviteUserArray));
+
     this.inviteUserForm = new FormGroup({
       firstName: new FormControl('', Validators.required),
       lastName: new FormControl('', Validators.required),
